Test website template
The folder has

Main html: the main page template which is suitable for the home page.
Has left menu on the side with hover elemnts, header area, body area and
footer area

grid.css: has style sheet for the grids. has weird code in there because
i was testing a lot of crap.